+++ src/GasPixel.java
@@ -25,7 +25,7 @@ public abstract class GasPixel extends Pixel {
 	 * @param b         The blue color value of the pixel
 	 * @param a         The alpha value of the pixel
 	 */
-	public GasPixel(double density, double temp, double condPoint, double depPoint, int r, int g, int b, int a) {
+	public GasPixel(double density, double temp, double condPoint, double depPoint, int r, int g, int b, double a) {
 		super(PixelType.GAS, density, temp, r, g, b, a);
 		this.condPoint = condPoint;
 		this.depPoint = depPoint;
