+++ src/SolidPixel.java
@@ -23,7 +23,7 @@ public abstract class SolidPixel extends Pixel {
 	 * @param a         The alpha value of the pixel
 	 */
 	SolidPixel(Pixel.PixelType type, double density, double temp, double meltPoint, double subPoint, int r, int g,
-			int b, int a) {
+			int b, double a) {
 		super(type, density, temp, r, g, b, a);
 		this.meltPoint = meltPoint;
 		this.subPoint = subPoint;
